This version I at least get a graph)

+++ ADALM-PLUTO/Modules/ble_hardware/pluto_tx.py
@@ -43,7 +43,7 @@ class PlutoTransmitter(Transmitter):
         '''
         self.sdr.tx_destroy_buffer()
         self.packet = packet
-        self.samples = createFSK(self.packet, 2 ** 14, self.ifreq, self.df, samples_per_bit=self.sample_rate * self.symbol_time, bit_time=self.symbol_time)
+        self.samples = createFSK(self.packet, 2 ** 14, self.ifreq, self.df, samples_per_bit= int(self.sample_rate * self.symbol_time), bit_time=self.symbol_time)
 
     def set_tx_power(self, power: float) -> None:
         '''
